// sanity/types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type SiteSettingsQueryResult = {
  siteTitle: string | null;
  siteDescription: string | null;
  logo: {
    asset: SanityImageAssetReference | null;
    hotspot: SanityImageHotspot | null;
    crop: SanityImageCrop | null;
    alt: string | null;
  } | null;
  logoText: {
    prefix: string | null;
    highlight: string | null;
    suffix: string | null;
  } | null;
  menuLabel: string | null;
  navMenus: Array<{
    _key: string;
    label: string | null;
    emptyText: string | null;
    links: Array<{
      _key: string;
      label: string | null;
      href: string | null;
    }> | null;
  }> | null;
  footerText: string | null;
} | null;